requirements.txt
django==5.2.10
djangorestframework==3.16.1
django_extensions==4.1
djangorestframework-simplejwt==5.5.1
django-filter==25.2
django-cors-headers
